Ensure that a small /RundownMaxMB value doesn't break rundown.
Rundown occurs in a series of steps.  Each step enables a provider that
can then produce events that flow into the session.  When a session is
size-constrained, as is the case with /RundownMaxMB, when the session
reaches is maximum size, the session is stopped.  Further attempts to
enable a provider will fail with the error `0x80071069: The instance name
passed was not recognized as valid by a WMI data provider.`  These
failures are expected, and are how handled by PerfView.