Fix NetTrace parsing for Start/Stop event names (#2186)
The work to add V6 NetTrace parsing regressed the special name handling for names that end with "Start" or "Stop".
Prior to this fix the suffix was correctly removed when the opcode in the trace was 0, but it was not removed
if the opcode reported in the file was Start or Stop.

The fixed code removes the "Start" and "Stop" suffixes when the opcode is either 0 or when it is already set to start/stop
respectively. An old V4 nettrace produced by the runtime would have opcode = 0 whereas a newer V5 nettrace from the
runtime would have the opcode set.

Ideally the runtime would be doing this event name transform on its own and the parser would not be in the business of
making these changes but for backwards compat we must.