Update NetTrace parameter parsing (#2200)
* Update NetTrace parameter parsing

BREAKING CHANGE: NetTrace V5 files that use the Array parameter type with character elements were previously marshalled into a .NET char[], but now TraceEvent converts them to a .NET System.String instead.

- Updated support for parameter types in NetTrace V6
  - Added new parameter types FixedLengthArray, DataLoc, RelLoc, and UTF8CodeUnit
  - Parameter type UTF16Char was renamed UTF16CodeUnit for in docs and code for clarity (chars are variable size in UTF16, code units are always 2 bytes). Behaviorly there is no change from before.
  - LengthPrefixedUTF8/16String were removed because they were duplicative with Array<UTF16CodeUnit> and Array<UTF8CodeUnit>. It didn't seem worthwhile to have a bunch of different string types when we could just interpret any of the 4 array variants with UTF8/16 elements as strings.

In order to add the new array types I needed to increase the functionality of the underlying PayloadFetch that TraceEvent uses to parse the parameters. PayloadFetch is serializable in the Etlx format so I wanted to added some tests that could round-trip the payload fetch through that serialization. That in turn led me to refactor TraceLog so that I could do it in-memory. I also found a variety of Debug asserts weren't true for the test data so I removed or relaxed them.

* Apply suggestions from code review

Co-authored-by: Brian Robbins <brianrob@microsoft.com>

* Update src/TraceEvent/DynamicTraceEventParser.cs

---------

Co-authored-by: Brian Robbins <brianrob@microsoft.com>