Implement PredefinedDynamicTraceEventParser For Known Dynamic Events (#2214)
* Implement a new parser type called PredefinedDynamicTraceEventParser that implements a
pre-compiled C# parser that targets dynamic events such as TraceLogging events.

The fundamental difference between this new type of parser and the existing
manifest-based pre-compiled parser is that the
PredefinedDynamicTraceEventParser does not have the same event identity
requirements as the manifest-based pre-compiled parser.  Instead of
requiring a durable provider ID and event ID, the
PreddefinedDynamicTraceEventParser requires a durable provider ID and
event name.  This enables users to build parsers for more flexible event
providers with the same parsing performance as manifest-based
pre-compiled parsers.

* TraceEventSamples needs to use the locally-built copy of TraceEvent to
ensure new APIs are present.

* GetTemplate --> TryGetTemplate