Special-Case BitMask Parsing
When an event contains a bitmask, it is sometimes described in metadata
as an array of pointers.  ETW differentiates between input and output
types, and so there is a possibility for metadata to specify the input
type as an array of UInt64 and the output type as an array of IntPtr.
Existing code can't handle this properly, so we must special case it.

The MSNT_SystemTrace/SystemConfig/ProcessorGroup event is an example of
one such event.